refactor: 拆分 useGameEngine.ts（538→324行）
- 新增 useSaveLoad.ts：存档读写 + 向后兼容迁移逻辑
- 新增 useCombatModal.ts：战斗弹窗状态 + 日志生成 + 死亡弹窗
- advanceTime 中重复死亡检测逻辑抽取为 runDeathCheck 辅助函数
- 类型通过 re-export 保持外部导入兼容
